requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
colorama==0.4.6
cssselect2==0.8.0
dataclasses-json==0.6.7
discord-py==2.6.4
distlib==0.4.0
distro==1.9.0
ebooklib==0.18
einops==0.8.1
et-xmlfile==2.0.0
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
google-auth==2.47.0
google-genai==1.57.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
identify==2.6.16
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-opentutorial==0.0.8
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
lxml==6.0.2
mammoth==1.11.0
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.1
markupsafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.1
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.5.1
portalocker==3.2.0
pre-commit==4.5.1
propcache==0.4.1
protobuf==6.33.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydyf==0.12.1
pypdfium2==4.30.0
pyphen==0.17.2
python-dotenv==1.2.1
python-pptx==1.0.2
pywin32==311
pyyaml==6.0.3
qdrant-client==1.16.2
rapidfuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
sqlalchemy==2.0.45
surya-ocr==0.17.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.13.0
virtualenv==20.36.1
wcwidth==0.2.14
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zopfli==0.4.0
zstandard==0.25.0
